cleanup gd headers
