+++ include/dolphin/gd.h
@@ -2,9 +2,13 @@
 #define _DOLPHIN_GD_H_
 
 #include <dolphin/gd/GDBase.h>
+#include <dolphin/gd/GDFile.h>
 #include <dolphin/gd/GDGeometry.h>
+#include <dolphin/gd/GDIndirect.h>
 #include <dolphin/gd/GDLight.h>
 #include <dolphin/gd/GDPixel.h>
 #include <dolphin/gd/GDTev.h>
+#include <dolphin/gd/GDTexture.h>
+#include <dolphin/gd/GDTransform.h>
 
 #endif
